Event state
There was a fundamental issue with the previous artwork-indexer code
when combined with its triggers (which weren't stored in this repo until
now): all of the event_queue INSERTS had the following clause in the
triggers:

    ON CONFLICT DO UPDATE SET attempts = 0

The idea being that this would reset the attempts on events that had
previously failed so that they could be retried, while having no effect
on events that haven't been run yet. The problem is, this would also
update events that were *already* running; in that case the correct
action is to queue a new event, because the running event may have
already fetched outdated metadata. But due to the UNIQUE index, this was
not possible.

This commit adds a `state` column to the event_queue table to track
whether an event is queued, running, failed, or completed. The UNIQUE
index now only applies to queued events, so triggers can safely "DO
NOTHING" "ON CONFLICT". Failed events are requeued by the indexer when
they fail, if necessary.

In order to maintain a history of events that may aid in debugging
issues in production, we also now store compeleted events for 90 days.
Failed events are stored indefinitely and must be handled by an admin.

As alluded to above, triggers and functions have been moved from
musicbrainz-server into this repo in order to ease maintenance and
logically group changes.