Kill async code
This commit replaces asyncpg with psycopg and aiohttp with requests.

artwork-indexer was designed from the start to take advantage of Python's
asyncio event loop, but that doesn't do much for us except allow us start
processing the next event while waiting for I/O on another event. Which is not
very useful, because we can safely (and easily) just run another
artwork-indexer process, and don't need any additional scalability given the
relatively low throughput of image uploads. The async model also increases the
complexity of the code, particularly with respect to error handling.