Make some queries resilient to connection failures
This commit adds a new class which wraps the psycopg connection, and, if that
connection disconnects somehow, is able to reconnect and retry the query that
failed (assuming `execute_with_retry` is used).

The motivation behind this is that there are certain queries related to
updating an event's state (i.e., marking it as completed or failed) that we
really do not want to fail, and would rather have the indexer retry forever
until any database issues are resolved.

There are other kinds of queries, like fetching the next event for processing,
which we don't need to retry in this way.  Since the indexer isn't doing
anything important, we can just let it crash and restart.