Fix concurrent event processing
Adding `autocommit=True` in e8b8623795a18bb9835de8f503b1b62cfa48e961 broke the
`SELECT ... FOR UPDATE ... SKIP LOCKED` check we have in place. We have to mark
the event as "running" in the same transaction as the `SELECT`.

I've added a concurrency test to ensure we don't regress here again.